Split ExtractNestedTypes tests into 3 sets of tests and ensure project is compilable (if not correct)
This is so I can start testing for expected behavior.
